Move EndpointEnumerator tests to separate module.
